utilities.c: add an iio_strlcpy, a "safer" version of strncpy
Add a private verison of strlcpy, from FreeBSD (Todd C. Miller).

strlcpy() is guarantee to NUL-terminate the resulting dest, even if it is
too small. Also note that strlcpy() and strlcat() only operate on true
'C' strings. This means that for strlcpy(), src must be NUL-terminated.

This is considered by many to be safer, and less error prone that strncpy,
which does not NUL-terminate on truncation.

Signed-off-by: Robin Getz <robin.getz@analog.com>